+++ beacon-chain/state/stateutil/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "hash_function.go",
         "helpers.go",
         "state_root.go",
+        "trie_helpers.go",
         "validators.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/state/stateutil",
@@ -37,13 +38,16 @@ go_test(
     srcs = [
         "state_root_cache_fuzz_test.go",
         "state_root_test.go",
+        "trie_helpers_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
         "//proto/beacon/p2p/v1:go_default_library",
+        "//shared/bytesutil:go_default_library",
         "//shared/featureconfig:go_default_library",
         "//shared/interop:go_default_library",
         "//shared/params:go_default_library",
+        "//shared/testutil:go_default_library",
         "@com_github_google_gofuzz//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
         "@com_github_prysmaticlabs_go_ssz//:go_default_library",
